ai-email-assistant/requirements.txt
# Web Framework
flask>=2.0.0
flask-cors>=3.0.0
flask-jwt-extended>=4.4.0

# Database
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0

# Machine Learning & NLP
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.0.0
numpy>=1.21.0
nltk>=3.8.0
spacy>=3.5.0
openai>=0.27.0

# Text Analysis & Processing
textblob>=0.17.0
langdetect>=1.0.9

# Security
bcrypt>=4.0.0
cryptography>=3.4.0
python-jose>=3.3.0
pyjwt>=2.3.0

# PDF Generation
fpdf2>=2.7.0  # Updated to fpdf2 for better features

# Environment & Configuration
python-dotenv>=0.19.0
environs>=9.5.0

# Email
smtplib-ssl>=1.0.1
email-validator>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0

# Development Tools
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
mypy>=1.0.0
pylint>=2.17.0

# Monitoring & Metrics 
prometheus-client>=0.16.0
statsd>=4.0.0

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0
